Move delete route

+++ back/routes/places.js
@@ -49,14 +49,19 @@ router.put('/:id', (request, response) => {
 
 router.delete('/:id', (req, res) => {
   const idPlace = req.params.id;
-  connection.query('DELETE FROM place WHERE id = ?', [idPlace], err => {
-    if (err) {
-      console.log(err);
-      res.status(500).send(`Erreur lors de la suppression d'un lieu, supprimer les évènements sur ${idPlace}`);
-    } else {
-      res.sendStatus(200);
-    }
-  });
-});
+  connection.query('SELECT happening_name, happening_picture, happening_category, happening_description, happening_date, happening_time, happening_time_end, happening_date_end, isItBookable, seats_bookable, local_photo, mapping FROM place INNER JOIN happening ON place.id= happening.place_id INNER JOIN admin ON place.admin_id=admin.id WHERE place.id = ? AND happening_date>=DATE(NOW()) OR happening_date IS NULL ORDER BY happening_date ASC', idPlace, (errGet, resultsGet) => {
+    if(resultsGet && resultsGet.length){
+      res.send("Merci de supprimer les évènements liés au préalable")
+    }else if(resultsGet.length===0){
+      connection.query('DELETE FROM place WHERE id = ?', [idPlace], err => {
+      if (err) {
+        console.log(err);
+        res.status(500).send(`Erreur lors de la suppression d'un lieu, supprimer les évènements sur ${idPlace}`);
+      } else {
+        res.sendStatus(200);
+      }
+    })}
+  })
+})
 
-  module.exports = router
+module.exports = router
